requirements.txt
flask
Jinja2
gunicorn
pylint
requests
pyyaml
